db-management/requirements.txt
PyYAML>=6.0.2
PyMySQL>=1.1.0
psycopg2-binary>=2.9.9
pymongo>=4.9.0
clickhouse-driver>=0.2.9
cryptography>=41.0.0
